+++ examples/async_ports.rs
@@ -21,8 +21,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let handle = Handle::current();
 
     let event_listener = EventListenerCfg::default()
-        .crossterm_input_listener(Duration::from_millis(10), 3)
         .with_handle(handle)
+        .async_crossterm_input_listener(Duration::default(), 3)
         .add_async_port(Box::new(AsyncPort::new()), Duration::from_millis(1000), 1);
 
     let mut app: Application<Id, Msg, UserEvent> = Application::init(event_listener);